feat(thetownoflight): first commit
* Add 2D LUT extrapolation (needed by The Town of Light mod)

Similar to the Starfield Luma implementation, but obviously simpler and more parametric.
I think this is a pretty good start, not sure if it's final though, but I can guarentee that it works and that the math should be correct.

* Add some generic color helpers and polish code

* Add support for The Town of Light.
The game outputs in gamma space with a paper white of 80 nits so it still needs an external post ReShade (e.g. Pumbo AdvancedAutoHDR) to be linearized properly. This can simply be fixed by adding a final RenoDX pass at the end on the swapchain passes.
Tonemapping is fixed to 1000 nits atm, it doesn't really need to go further for this game.