Add Bioshock (Remastered)
Upgrading textures and removing a saturate() was all I had to do. The game was clipped, there was no tonemapping (and there still isn't).
Bioshock 2 is probably the same.

Some of the render target textures are still not upgraded and are still R11G11B10F, unclear why.

Possible issues:
- Shadow maps are broken?
- Some UI blends get too dark? Like the bottom right radio talk picture